+++ easychat/lib/easychat.dart
@@ -14,7 +14,6 @@ export './src/chat.exception.dart';
 
 // Widgets ====================================================
 export './src/widgets/chat.bubble.dart';
-export './src/widgets/chat.bubble.long_press.popup_menu.dart';
 export './src/widgets/chat.bubble.reply.dart';
 export './src/widgets/chat.messages.list_view.dart';
 export './src/widgets/chat.room.input_box.dart';
