+++ easychat/lib/src/chat.service.dart
@@ -200,14 +200,12 @@ class ChatService {
   Future<void> showEditMessageDialog(
     BuildContext context, {
     required ChatMessage message,
-    required ChatRoom room,
   }) async {
     await showDialog(
       context: context,
       builder: (context) {
         return EditChatMessageDialog(
           message: message,
-          room: room,
         );
       },
     );
